backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-multipart
python-dotenv
psycopg2-binary
requests
blynk-library-python
numpy
scipy
pandas
